Sets up the airplane integration test to run in any environment
