+++ pom.xml
@@ -14,7 +14,6 @@
         <junit.platform.version>1.0.1</junit.platform.version>
         <hamcrest-junit.version>2.0.0.0</hamcrest-junit.version>
         <sqlite-jdbc.version>3.20.1</sqlite-jdbc.version>
-        <maven-failsafe-plugin.version>2.13</maven-failsafe-plugin.version>
     </properties>
     <build>
         <plugins>
@@ -32,25 +31,6 @@
                         <version>${junit.platform.version}</version>
                     </dependency>
                 </dependencies>
-            </plugin>
-            <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-failsafe-plugin</artifactId>
-                <version>${maven-failsafe-plugin.version}</version>
-                <executions>
-                    <execution>
-                        <id>integration-test</id>
-                        <goals>
-                            <goal>integration-test</goal>
-                        </goals>
-                    </execution>
-                    <execution>
-                        <id>verify</id>
-                        <goals>
-                            <goal>verify</goal>
-                        </goals>
-                    </execution>
-                </executions>
             </plugin>
         </plugins>
     </build>
